+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Spend in single transaction coins owned by multiple wallets (same altcoin plugin)
 - Added logger for the app and proper error handling
 - Built-in support for SkyWallet hardware wallet as signer for Skycoin transactions
 - Built-in support iand GUI for applying operations upon SkyWallet hardware devices